Who needs to save iops in a tic tac toe game?
So... I realized the code wasn't actually as functional/scalable as I thought. Diagonals were only being checked if they were exactly down the middle.
So, I made ugly code that is going to hurt the brain of whoever tries to read it... It's not TERRIBLE, but... it is brute force-y.
A better solution would be to only check based on the play made, but I'm not sure how I'd do that without making another array with diagonals listed. I could do that, but I feel like that would make the same amount of calls, if not a couple extra for making the list of diagonals.
Listen... I'm not super proud of it, but it's 3:03 am and I don't have the brain power to consider implementing a way that this could be less brute force-y.