Cleaned up flow of code and added some comments
`find_winner()` wasn't working, due to me being bad and assuming a `while not winner:` is going to check `winner` at every step, instead of just at the start. Fixed that. :/
Added comments to the `check_all_diagonals_X()` functions, so that it's more clear what they're all doing, as they all piggy-back off of the first one and it's hard to mentally keep track of which does what in case there are issues.
Removed `TODO:` block from code, I'll leave that stuff in README.md.